backend/requirements.txt
django>=4.2
djangorestframework>=3.14
PyJWT>=2.8
supabase>=2.0
python-dotenv>=1.0
requests>=2.31
psycopg2-binary>=2.9
django-cors-headers>=4.3
